feat(diffusion_planner): add ego state selection and interpolation utilities
